+++ packages/sdk/CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## Next
 
+### Added
+
+- Added optional `path` parameter to `processProtectedData` method
+
 ### Changed
 
-- Add optional `path` param in `processProtectedData`
 - Moved `getResultFromCompletedTask` method from DataProtectorSharing module (DPS) to DataProtectorCore module (DPC).
 
 ## [2.0.0-beta.11] (2025-01-29)